Protected event routes requiring user auth

+++ controllers/events.js
@@ -37,6 +37,8 @@ exports.getEvent = AsyncHandler(async (req, res, next) => {
 // @route   POST /events
 // @access  Public
 exports.addEvent = AsyncHandler(async (req, res, next) => {
+  req.body.user = req.user;
+  console.log(req.user);
   const event = new Event(req.body);
   await event.save();
 
@@ -46,29 +48,40 @@ exports.addEvent = AsyncHandler(async (req, res, next) => {
 // @desc    Delete an event
 // @route   DELETE /events/:id
 // @access  Public
-exports.deleteEvent = async (req, res, next) => {
-  try {
-    const event = await Event.findById(req.params.id);
+exports.deleteEvent = AsyncHandler(async (req, res, next) => {
+  const event = await Event.findById(req.params.id);
 
-    if (!event) {
-      next(new ErrorResponse(`Event not found with ID ${req.params.id}`, 404));
-    }
-    await Event.deleteOne(event);
-
-    return res.status(204).json({});
-  } catch (err) {
-    next(err);
+  if (!event) {
+    return next(
+      new ErrorResponse(`Event not found with ID ${req.params.id}`, 404)
+    );
   }
-};
+
+  console.log(event.user, req.user);
+
+  if (event.user.toString() !== req.user) {
+    return next(
+      new ErrorResponse("You are not authorized to delete this event", 403)
+    );
+  }
+
+  await Event.deleteOne(event);
+
+  return res.status(204).json({});
+});
 
 // @desc    Update an event
 // @route   PUT /events/:id
 // @access  Public
-exports.updateEvent = AsyncHandler(async (req, res) => {
+exports.updateEvent = AsyncHandler(async (req, res, next) => {
   let event = await Event.findById(req.params.id);
 
   if (!event) {
     next(new ErrorResponse(`Event not found with ID ${req.params.id}`, 404));
+  }
+
+  if (event.user.toString() !== req.user) {
+    next(new ErrorResponse("You are not authorized to edit this event", 403));
   }
 
   if (
